added confirmation modal before delete, fixed pre-populated data in edit borrowed list

+++ src/components/AuthorList.jsx
@@ -73,6 +73,11 @@ const AuthorList = () => {
   };
 
   const handleDelete = async (id) => {
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this author?"
+    );
+    if (!confirmDelete) return;
+
     try {
       const response = await fetch(`${apiUrl}/authors/${id}`, {
         method: "DELETE",
